Changed query names to make more sense in signalfx
