+++ rubrik_polaris/lib/common/graphql/mutation_accounts_aws_delete_commit.graphql
