+++ rubrik_polaris/lib/common/graphql/mutation_instances_azure_restore.graphql
