+++ rubrik_polaris/lib/common/graphql/mutation_instances_ec2_restore.graphql
