+++ rubrik_polaris/lib/common/graphql/mutation_instances_gce_restore.graphql
