+++ rubrik_polaris/lib/common/graphql/query_core_snappable_snapshots.graphql
