+++ rubrik_polaris/lib/common/graphql/query_instances_aws_ec2.graphql
