+++ rubrik_polaris/lib/common/graphql/query_instances_azure_iaas.graphql
