+++ rubrik_polaris/lib/common/graphql/query_instances_gcp_gce.graphql
