+++ rubrik_polaris/lib/common/graphql/query_instances_vmware_vsphere.graphql
