+++ rubrik_polaris/lib/common/graphql/query_storage_aws_ebs.graphql
