+++ rubrik_polaris/rubrik_polaris.py
@@ -27,8 +27,8 @@ import pprint
 
 class PolarisClient:
     # Public
-    from .lib.common.polaris import get_sla_domains, submit_on_demand, submit_assign_sla, get_task_status
-    from .lib.common.polaris import get_snapshots
+    from .lib.common.core import get_sla_domains, submit_on_demand, submit_assign_sla, get_task_status
+    from .lib.common.core import get_snapshots
     from .lib.accounts import get_accounts_azure, get_accounts_gcp
     from .lib.accounts import get_accounts_aws, get_accounts_aws_detail, get_account_aws_native_id
     from .lib.accounts import add_account_aws, delete_account_aws
